requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==4.2.0
black==23.11.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.7
django-celery-beat==2.5.0
django-rest-framework==0.1.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
frozenlist==1.4.0
idna==3.6
importlib-resources==6.1.1
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kombu==5.3.4
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
prompt-toolkit==3.0.41
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.31.1
requests==2.31.0
rpds-py==0.13.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
yarl==1.9.3
zipp==3.17.0
